+++ project/src/components/film-tabs/film-tabs.jsx
@@ -1,4 +1,5 @@
-import React, {useState} from 'react';
+import React from 'react';
+import {Link, useParams} from 'react-router-dom';
 import filmProp from '../film-screen/film.prop';
 import FilmOverview from '../film-overview/film-overview';
 import FilmDetails from '../film-details/film-details';
@@ -6,19 +7,11 @@ import FilmReviews from '../film-reviews/film-reviews';
 import {Tabs} from '../../const';
 
 function FilmTabs(props) {
-  const [activeTab, setActiveTab] = useState(Tabs.OVERVIEW);
   const {film} = props;
-
-  function handleActiveTabChange(evt) {
-    if (evt.target.tagName === 'A') {
-      evt.preventDefault();
-
-      setActiveTab(evt.target.textContent.toLowerCase());
-    }
-  }
+  const {tab = Tabs.OVERVIEW} = useParams();
 
   function getComponentByActiveTab() {
-    switch (activeTab) {
+    switch (tab) {
       case Tabs.DETAILS:
         return <FilmDetails film={film} />;
       case Tabs.REVIEWS:
@@ -30,16 +23,16 @@ function FilmTabs(props) {
 
   return (
     <div className="film-card__desc">
-      <nav className="film-nav film-card__nav" onClick={handleActiveTabChange}>
+      <nav className="film-nav film-card__nav">
         <ul className="film-nav__list">
-          <li className={`film-nav__item ${activeTab === Tabs.OVERVIEW ? 'film-nav__item--active' : ''}`}>
-            <a href="##" className="film-nav__link">Overview</a>
+          <li className={`film-nav__item ${tab === Tabs.OVERVIEW ? 'film-nav__item--active' : ''}`}>
+            <Link to={`/films/${film.id}`} className="film-nav__link">Overview</Link>
           </li>
-          <li className={`film-nav__item ${activeTab === Tabs.DETAILS ? 'film-nav__item--active' : ''}`}>
-            <a href="##" className="film-nav__link">Details</a>
+          <li className={`film-nav__item ${tab === Tabs.DETAILS ? 'film-nav__item--active' : ''}`}>
+            <Link to={`/films/${film.id}/details`} className="film-nav__link">Details</Link>
           </li>
-          <li className={`film-nav__item ${activeTab === Tabs.REVIEWS ? 'film-nav__item--active' : ''}`}>
-            <a href="##" className="film-nav__link">Reviews</a>
+          <li className={`film-nav__item ${tab === Tabs.REVIEWS ? 'film-nav__item--active' : ''}`}>
+            <Link to={`/films/${film.id}/reviews`} className="film-nav__link">Reviews</Link>
           </li>
         </ul>
       </nav>
