+++ project/src/components/player-screen/player-screen.jsx
@@ -2,22 +2,12 @@ import React from 'react';
 import {AppRoute} from '../../const';
 import {useHistory, useParams} from 'react-router-dom';
 import films from '../../mocks/films';
+import {formatRunTimeForPlayer} from '../../utils';
 
 function PlayerScreen() {
   const history = useHistory();
   const {id} = useParams();
   const film = films.find((currentFilm) => currentFilm.id === parseInt(id, 10));
-
-  function formatRunTime(runTime) {
-    const hours = Math.floor(runTime / 60);
-    const minutes = runTime % 60 > 9 ? runTime % 60 : `0${runTime % 60}`;
-
-    if (hours === 0) {
-      return `${minutes}:00`;
-    }
-
-    return `${hours}:${minutes}:00`;
-  }
 
   return (
     <div className="player">
@@ -31,7 +21,7 @@ function PlayerScreen() {
             <progress className="player__progress" value="0" max="100"></progress>
             <div className="player__toggler">Toggler</div>
           </div>
-          <div className="player__time-value">{formatRunTime(film.runTime)}</div>
+          <div className="player__time-value">{formatRunTimeForPlayer(film.runTime)}</div>
         </div>
 
         <div className="player__controls-row">
