+++ project/src/const.js
@@ -3,6 +3,8 @@ export const AppRoute = {
   SIGN_IN: '/login',
   MY_LIST: '/mylist',
   FILM: '/films/:id',
+  FILM_DETAILS: '/films/:id/:tab',
+  FILM_REVIEWS: '/films/:id/:tab',
   ADD_REVIEW: '/films/:id/review',
   PLAYER: '/player/:id',
 };
